+++ app/server-utils/stampy.ts
@@ -1,4 +1,5 @@
 import {withCache} from '~/server-utils/kv-cache'
+import { urlToIframe } from './url-to-iframe'
 import MarkdownIt from 'markdown-it'
 import MarkdownItFootnote from 'markdown-it-footnote'
 
@@ -186,28 +187,6 @@ const convertToQuestion = (title: string, v: CodaRow['values']): Question => ({
       : [],
   status: v['Status']?.name as QuestionStatus,
 })
-
-/**
- * Replaces all of the URLs in text with iframes of the URLs
- * @param text text in which URLs should be replaced
- */
-const urlToIframe = (text: string): string => {
-  const whitelistedHosts = [
-    "coda.io",
-    "airtable.com", 
-    "aisafety.world"
-  ]
-  // Regex is from: https://stackoverflow.com/a/3809435
-  const httpsUrlRegex = new RegExp(/https?:\/\/(www\.)?[-a-zA-Z0-9@:%._+~#=]{1,256}\.[a-zA-Z0-9()]{1,6}\b([-a-zA-Z0-9()@:%_+.~#?&//=]*)?/gi)
-  const matches = text.matchAll(httpsUrlRegex)
-  let updatedText = text
-  for (const match of matches) {
-    const url = match[0] 
-    if (whitelistedHosts.includes(new URL(url).host))
-    updatedText = text.replace(url, `<iframe src="${url}"/>`)
-  }
-  return updatedText
-}
 
 export const loadQuestionDetail = withCache('questionDetail', async (question: string) => {
   const rows = await getCodaRows(
